+++ packages/tempo/vitest.config.ts
@@ -44,7 +44,6 @@ export default defineConfig({
       { find: /^#tempo\/plugin\/plugin\.(.*)\.js$/, replacement: resolve(__dirname, './dist/plugin/plugin.$1.js') },
       { find: /^#tempo\/plugin\/extend\/(.*)\.js$/, replacement: resolve(__dirname, './dist/plugin/extend/$1.js') },
       { find: /^#tempo\/engine\/(.*)\.js$/, replacement: resolve(__dirname, './dist/engine/$1.js') },
-      { find: /^#tempo\/parse\/(.*)\.js$/, replacement: resolve(__dirname, './dist/parse/$1.js') },
       { find: /^#tempo\/module\/(.*)\.js$/, replacement: resolve(__dirname, './dist/module/$1.js') },
       { find: /^#tempo\/plugin\/term\/(.*)\.js$/, replacement: resolve(__dirname, './dist/plugin/term/$1.js') },
       { find: /^#tempo\/support$/, replacement: resolve(__dirname, './dist/support/support.index.js') },
@@ -65,7 +64,6 @@ export default defineConfig({
       { find: /^#tempo\/plugin\/plugin\.(.*)\.js$/, replacement: resolve(__dirname, './src/plugin/plugin.$1.ts') },
       { find: /^#tempo\/plugin\/extend\/(.*)\.js$/, replacement: resolve(__dirname, './src/plugin/extend/$1.ts') },
       { find: /^#tempo\/engine\/(.*)\.js$/, replacement: resolve(__dirname, './src/engine/$1.ts') },
-      { find: /^#tempo\/parse\/(.*)\.js$/, replacement: resolve(__dirname, './src/parse/$1.ts') },
       { find: /^#tempo\/module\/(.*)\.js$/, replacement: resolve(__dirname, './src/module/$1.ts') },
       { find: /^#tempo\/plugin\/term\/(.*)\.js$/, replacement: resolve(__dirname, './src/plugin/term/$1.ts') },
       { find: /^#tempo\/support$/, replacement: resolve(__dirname, './src/support/support.index.ts') },
